[FIX] notation
