+++ wepelo_equipment/__manifest__.py
@@ -13,10 +13,10 @@
         'security/ir.model.access.csv',
         'data/wepelo_mail_data.xml',
         'views/assets.xml',
-        'views/maintenance_equipment_category_views.xml',
-        'views/wepelo_maintenance_views.xml',
+        'views/sm_maintenance_equipment_category_views.xml',
+        'views/sm_views.xml',
         'wizard/mail_activity_edit_wizard_views.xml',
-        'views/mail_activity_views.xml',
+        'views/sm_mail_activity_views.xml',
         'report/begehung_safety_master.xml',
         'report/wepelo_maintenance_equipment_strichcode.xml',
         'report/folgebegehung_safety_master.xml',
@@ -24,12 +24,11 @@
         'report/protocol_betriebsanweisung_template.xml',
         'report/gefahrstoff_verzeichnis.xml',
         'report/unterweisung.xml',
-        'views/menu.xml',
-        'views/wepelo_equipment_protocol_views.xml',
-        'views/wepelo_repair_view.xml',
-        'views/wepelo_customer_view.xml',
-        'views/equipment_portal_templates.xml',
-        'views/res_users_views.xml',
+        'views/sm_menu.xml',
+        'views/sm_protocol_views.xml',
+        'views/sm_customer_view.xml',
+        'views/sm_equipment_portal_templates.xml',
+        'views/sm_res_users_views.xml',
     ],
     'demo': [],
     'installable': True,
