+++ wepelo_equipment/models/.ipynb_checkpoints/sm_a_gefahrstoff_verzeichnis-checkpoint.py
