+++ wepelo_equipment/models/.ipynb_checkpoints/sm_customer-checkpoint.py
