+++ wepelo_equipment/models/.ipynb_checkpoints/sm_history-checkpoint.py
