+++ wepelo_equipment/models/.ipynb_checkpoints/sm_ir_attachment-checkpoint.py
