+++ wepelo_equipment/models/.ipynb_checkpoints/sm_mail_activity-checkpoint.py
