+++ wepelo_equipment/models/.ipynb_checkpoints/sm_protocol-checkpoint.py
