+++ wepelo_equipment/models/.ipynb_checkpoints/sm_qr_code-checkpoint.py
