+++ wepelo_equipment/models/.ipynb_checkpoints/sm_res_users-checkpoint.py
