+++ wepelo_equipment/models/.ipynb_checkpoints/sm_system-checkpoint.py
