+++ wepelo_equipment/models/sm_ir_attachment.py
