+++ wepelo_equipment/models/sm_mail_activity.py
