+++ wepelo_equipment/models/sm_maintenance.py
@@ -51,7 +51,7 @@ class MaintenanceEquipment(models.Model):
     attachment_count = fields.Integer(
         string="Attachments Number", compute="_compute_attachment_count"
     )
-    name_seq = fields.Char(string="Nummer", default=lambda self: self._get_next_serial_no_name(),  store=True, readonly=True)
+    name_seq = fields.Char(string="Nummer",  store=True, readonly=True)
     ir_sequence_id = fields.Many2one('ir.sequence')
 
     def attachment_tree_view(self):
