+++ wepelo_equipment/models/sm_qr_code.py
