+++ wepelo_equipment/models/sm_system.py
