+++ wepelo_equipment/security/.ipynb_checkpoints/ir.model.access-checkpoint.csv
@@ -13,7 +13,6 @@ access_maintenance_equipment_portal,access_maintenance_equipment_portal,model_ma
 access_equipment_protocol_portal,access_equipment_protocol_portal,model_equipment_protocol,base.group_portal,1,1,0,0
 access_maintenance_equipment_category_portal,access_maintenance_equipment_category_portal,maintenance.model_maintenance_equipment_category,base.group_portal,1,0,0,0
 access_mail_activity_portal,access_mail_activity_portal,mail.model_mail_activity,base.group_portal,1,1,1,0
-access_maintenance_request_portal,access_maintenance_request_portal,wepelo_equipment.model_maintenance_request,base.group_portal,1,1,1,0
 access_maintenance_stage_portal,access_maintenance_stage_portal,maintenance.model_maintenance_stage,base.group_portal,1,0,0,0
 access_maintenance_team_portal,access_maintenance_team_portal,maintenance.model_maintenance_team,base.group_portal,1,0,0,0
 access_res_users_portal,access_res_users_portal,base.model_res_users,base.group_portal,1,0,0,0
