+++ wepelo_equipment/views/.ipynb_checkpoints/sm_customer_view-checkpoint.xml
