+++ wepelo_equipment/views/.ipynb_checkpoints/sm_equipment_portal_templates-checkpoint.xml
