+++ wepelo_equipment/views/.ipynb_checkpoints/sm_mail_activity_views-checkpoint.xml
