+++ wepelo_equipment/views/.ipynb_checkpoints/sm_maintenance_equipment_category_views-checkpoint.xml
