+++ wepelo_equipment/views/.ipynb_checkpoints/sm_menu-checkpoint.xml
