+++ wepelo_equipment/views/.ipynb_checkpoints/sm_protocol_views-checkpoint.xml
