+++ wepelo_equipment/views/.ipynb_checkpoints/sm_res_users_views-checkpoint.xml
