+++ wepelo_equipment/views/.ipynb_checkpoints/sm_views-checkpoint.xml
