+++ wepelo_equipment/views/sm_customer_view.xml
