+++ wepelo_equipment/views/sm_equipment_portal_templates.xml
