+++ wepelo_equipment/views/sm_mail_activity_views.xml
