+++ wepelo_equipment/views/sm_maintenance_equipment_category_views.xml
