+++ wepelo_equipment/views/sm_menu.xml
