+++ wepelo_equipment/views/sm_protocol_views.xml
