+++ wepelo_equipment/views/sm_views.xml
